Removed typing in generic classifier.

+++ src/n-dim/types.jl
@@ -9,7 +9,7 @@ A mutable struct representing a point in D-dimensional space.
 @param coords::SVector{D,ComplexF64} The coordinates of the point in a Riemann chart.
 @param active::Bool Whether the point is active or not. (Criteria depends on parameters given initially.)
 """
-mutable struct Point{D::Int}
+mutable struct Point{D}
     coords::SVector{D,ComplexF64}
     active::Bool
 end
@@ -25,7 +25,7 @@ A mutable struct representing a simplex in D-dimensional space.
 @param vertices::SVector{D,Point} The vertices of the simplex.
 @param active::Bool Whether the simplex is active or not.
 """
-mutable struct Simplex{D::Int}
+mutable struct Simplex{D}
     vertices::SVector{D,Point}
     active::Bool
 end
